# docs/api-contract.yaml
openapi: 3.0.3
info:
  title: Hypatia Backend API
  description: >
    Contract between the Flutter frontend and the FastAPI backend for the
    Hypatia LLM wiki application. Covers health, Classes CRUD, file
    upload/serve/delete, wiki tree/page CRUD/search/export/lint/rebuild,
    chat WebSocket + history, task status, and class backup/import.
  version: "0.4.0"
servers:
  - url: http://127.0.0.1:8000
    description: >-
      Local dev backend. Port may vary 8000-8010 if 8000 is taken
      (see Task 1.9 auto-launch/port management).
paths:
...
  /api/classes:
    get:
      summary: List all Classes
      operationId: listClasses
      responses:
        "200":
          description: All Classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassRead"
    post:
      summary: Create a new Class
      operationId: createClass
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassCreate"
      responses:
        "201":
          description: Class created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassRead"
...
  /api/classes/{class_id}/backup:
    parameters:
      - $ref: "#/components/parameters/ClassId"
    post:
      summary: Export a Class as a backup ZIP
      operationId: backupClass
      responses:
        "200":
          description: Backup ZIP file
          content:
            application/zip: {}
        "404":
          $ref: "#/components/responses/NotFound"
  /api/classes/import:
    post:
      summary: Import a Class from a backup ZIP
      operationId: importClass
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        "201":
          description: Class imported successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  parameters:
    ClassId:
      name: class_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
        code:
          type: string
      required: [detail, code]
    ClassCreate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
      required: [name]
...
    ClassRead:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name, description, created_at, updated_at]
...
    ImportResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        file_count:
          type: integer
        page_count:
          type: integer
        message_count:
          type: integer
      required: [id, name, file_count, page_count, message_count]
...